requirements-production.txt
MySQL-python==1.2.4
gunicorn==0.17.2
boto==2.1.1
django-storages==1.1.4
django-ses==0.4.1
python-dateutil==1.5
